requirements.txt
jupyter
matplotlib
numpy
torch
torchvision
Pillow
tqdm
prettytable
pytorch